apps/healthcare_appeals_decisioning/requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-groq>=0.1.2
pypdf>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
